# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for NeuralVault",
        "title": "NeuralVault API",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/query": {
            "post": {
                "description": "Embeds the question from a JSON body {\"workspace_id\": uuid, \"question\": string, \"top_k\": int, \"conversation_id\": uuid (optional)} and returns the top-k matching chunks ordered by descending similarity score. If conversation_id is set, persists the question as a message on that conversation (no answer is generated here to persist).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "query"
                ],
                "summary": "Query a workspace",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/query/stream": {
            "post": {
                "description": "Runs retrieval for a JSON body {\"workspace_id\": uuid, \"question\": string, \"top_k\": int, \"conversation_id\": uuid (optional)}, then streams SSE events: one \"sources\" event with the grounding chunks, incremental \"token\" events, and a terminal \"done\" or \"error\" event. If conversation_id is set, persists the question and the completed answer (with sources) as messages on that conversation.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "query"
                ],
                "summary": "Stream a grounded answer (SSE)",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
...
}